all-done

+++ services/employee.service.ts
@@ -1,14 +1,16 @@
 import EmployeeRepository from "../repositories/employee.repositories";
-import Employee, { EmployeeRole } from "../entities/employee.entity";
+import Employee, { EmployeeRole, EmployeeStatus } from "../entities/employee.entity";
 import Address from "../entities/address.entity";
 import bcrypt from "bcrypt";
 import { LoggerService } from "./logger.service";
+import Department from "../entities/department.entity";
+import HttpException from "../exceptions/httpException";
 
 class EmployeeService{
-     private logger=LoggerService.getInstance(EmployeeService.name);
+     private logger=LoggerService.getInstance('EmployeeService');
     constructor(private employeeRepository:EmployeeRepository){}
 
-    async createEmployee(email:string,name:string,age:number,address:Address,password:string,role:EmployeeRole):Promise<Employee>{
+    async createEmployee(email:string,name:string,age:number,address:Address,password:string,role:EmployeeRole,department:Department,status:EmployeeStatus,experience:number,joiningdate:Date):Promise<Employee>{
         const newEmployee=new Employee();
         newEmployee.name=name;
         newEmployee.email=email;
@@ -16,6 +18,10 @@ class EmployeeService{
         newEmployee.address=address;
         newEmployee.password=await bcrypt.hash(password,10);
         newEmployee.role=role;
+        newEmployee.department=department;
+        newEmployee.status=status;
+        newEmployee.Experience=experience;
+        newEmployee.joiningdate=joiningdate;
         return this.employeeRepository.create(newEmployee);
     }
 
@@ -27,6 +33,7 @@ class EmployeeService{
 
    async getEmployeeByID(id:number):Promise<Employee>{
         let employee=this.employeeRepository.findOneByID(id);
+        this.logger.info(employee);
         if(!employee){
           throw new Error("Employee Not Found");
         }
@@ -37,24 +44,32 @@ class EmployeeService{
         return this.employeeRepository.findOneByEmail(email);
    }
 
-   async updateEmployee(id:number,name:string,email:string):Promise<void>{
+   async updateEmployee(id:number,name:string,email:string,age:number,address:Address,password:string,role:EmployeeRole,department:Department,status:EmployeeStatus,joiningdate:Date,experience:number):Promise<void>{
         const existingEmployee=this.employeeRepository.findOneByID(id);
         if(existingEmployee){
             const employee=new Employee();
             employee.name=name;
             employee.email=email;
+            employee.age=age;
+            employee.address=address;
+            employee.password=password;
+            employee.role=role;
+            employee.department=department;
+            employee.status=status;
+            employee.joiningdate=joiningdate;
+            employee.Experience=experience;
             await this.employeeRepository.update(id,employee);
         }
    }
 
    async deleteEmployeeByID(id:number):Promise<void>{
-     const employee=await this.employeeRepository.findOneByID(id);
-    await this.employeeRepository.deleteOneByID(id);
+      const employee=await this.employeeRepository.findOneByID(id);
+      if(!employee){
+          throw new HttpException(404,"employee not found");
+      }
+     await this.employeeRepository.deleteOneByID(id);
    }
 
-   async remove(empid:number):Promise<void>{
-    await this.employeeRepository.remove(empid);
-   }
 }
 
 export default EmployeeService;
